requirements/production.txt
# ==================================================
# PRODUCTION REQUIREMENTS
# ==================================================

-r base.txt

# Production server
gunicorn>=23.0.0

# AWS S3 for media storage
boto3>=1.40.0
django-storages>=1.14.2

# Performance and caching
redis>=7.0.0
django-redis>=6.0.0
hiredis>=3.3.0

# Monitoring and error tracking
sentry-sdk>=2.42.0

